Fix errors importing the Tor Project's OpenPGP public key.
Attempting to import the Tor Project's Debian archive signing key
reliably produces "packet too large" errors, the same as that described
at the bug here:

    https://bitbucket.org/skskeyserver/sks-keyserver/issues/57/anyone-can-make-any-pgp-key-unimportable

I'm not certain if this is a result of someone "making the Tor Project's
GPG key unimportable" as a DoS attack or if it's some long-standing
issue with the keyserver infrastructure but, either way, a simple
workaround is to provide a known-good key as part of the role itself.

This commit also takes the opportunity to clean the style on some tasks.